Drop AclService interface as its redundant
- Move dryrun control to AclEnforcer
- Add package-info docs for mixins

+++ kafka_acl_enforcer/src/main/java/com/tesla/data/acl/AclEnforcer.java
@@ -13,19 +13,25 @@ import java.util.List;
 
 public class AclEnforcer extends Enforcer<AclBinding> {
   private final AclService aclService;
+  private final boolean dryRun;
 
-  public AclEnforcer(Collection<AclBinding> configured, AclService aclService, boolean safemode) {
+  public AclEnforcer(Collection<AclBinding> configured, AclService aclService, boolean safemode, boolean dryRun) {
     super(configured, aclService::listExisting, AclBinding::equals, safemode);
     this.aclService = aclService;
+    this.dryRun = dryRun;
   }
 
   @Override
   protected void create(List<AclBinding> toCreate) {
-    aclService.create(toCreate);
+    if (!dryRun) {
+      aclService.create(toCreate);
+    }
   }
 
   @Override
   protected void delete(List<AclBinding> toDelete) {
-    aclService.delete(toDelete);
+    if (!dryRun) {
+      aclService.delete(toDelete);
+    }
   }
 }
